requirements.txt
Cython==0.17.1
Django==1.4.2
EasyProcess==0.1.4
Fabric==1.4.3
South==0.7.6
Unidecode==0.04.9
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.17
boto==2.6.0
celery==3.0.11
discogs-client==1.1.1
dj-database-url==0.2.1
django-adminplus==0.1.7
django-appconf==0.5
django-celery==3.0.11
django-compressor==1.2
django-debug-toolbar==0.9.4
django-jsonfield==0.8.11
django-nose==1.1
django-ses==0.4.1
django-social-auth==0.7.7
django-uuidfield==0.4.0
gunicorn==0.15.0
httplib2==0.7.6
ipdb==0.7
ipython==0.13.1
kombu==2.4.7
lxml==3.0.1
markdown2==2.1.0
nose==1.2.1
oauth2==1.5.211
pika==0.9.5
psycopg2==2.4.5
pycrypto==2.6
python-dateutil==1.5
python-openid==2.2.5
pytz==2012g
requests==0.14.2
ssh==1.7.14
wsgiref==0.1.2
